backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic
python-dotenv
